change datatype of capacity to smallint

+++ database-sql/dbo/StoredProcedures/event_Update.sql
@@ -1,7 +1,7 @@
 CREATE OR ALTER PROCEDURE event_Update
 (@id INT,@eventName NVARCHAR(100),
 @eventDateTime DATETIME2(7),@hall NVARCHAR(50),
-@capacity TINYINT,@isPrivate BIT,@targetAudience SMALLINT=null)
+@capacity SMALLINT,@isPrivate BIT,@targetAudience SMALLINT=null)
 AS
 UPDATE event
 SET Name = COALESCE(@eventName,Name),
